Pass extra querystring parameters to provider
Querystring parameters which are not reserved fields or known provider
fields are now passed to the provider as `extra_params`.

This is not a breaking change since providers are expected to accept
**kwargs.

Currently this is only used for the oracle provider, but it's available
for interpreting custom parameters for any provider.